+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## [0.1.0-SNAPSHOT]
+## [0.1.1]
 
-Version created from `Clojure2d` library.
+### Added
+
+* BesselJ
+* all Gamma function variants (regularized, etc)
+* regularized Beta
+* log1p
+* `estimation-strategy` can be passed to any function calculating quantiles/percentiles
+* ->seq can accept number of samples now
+* low-exp, high-exp to find lower/greater exponent for given base and number
+
+### Changed
+
+* `stats-map` doesn't contain select keys option anymore
+
+## [0.1.0]
+
+Initial version created from `Clojure2d` library.